refacto contactForm
